common lib
