+++ common_lib/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "common_lib"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+serde = "1.0.179"
